Extract order duplication to a base
For of our certificate order are repeating the same behavior to
place a new order, so this commit extract those common behavior
to a base class and they just define the type on their own and
the base class automatically usages that to create a proper end
point for each type of the certificate.
