+++ spec/digicert/order/ssl_plus_spec.rb
@@ -3,7 +3,7 @@ require "spec_helper"
 RSpec.describe Digicert::Order::SSLPlus do
   describe ".create" do
     it "creates a new order for ssl plus certificate" do
-      stub_digicert_ssl_plus_create(order_attributes)
+      stub_digicert_order_create_api("ssl_plus", order_attributes)
       order = Digicert::Order::SSLPlus.create(order_attributes)
 
       expect(order.id).not_to be_nil
