--- spec/fixtures/ssl_wildcard_order_created.json
@@ -1,9 +1,0 @@
-{
-  "id": 542772,
-  "requests": [
-    {
-      "id": 922432,
-      "status": "pending"
-    }
-  ]
-}
